+++ myokit/formats/cellml/v2/__init__.py
@@ -13,7 +13,6 @@ from ._api import (     # noqa
     Model,
     Units,
     Variable,
-    is_basic_real_number_string,
     is_identifier,
     is_integer_string,
     is_real_number_string,